Move plot() to tools.py
